galaxyxml_analysis/requirements.txt
bioblend
python-hglib
lxml
pandas
matplotlib
bibtexparser

